refactor: change the series library to be named library
Having the library and feature both named series causes some compilation issues, it also confuses
parts in Gradle. For simplicities sake, change the name of the library part to just "library"
