+++ app/src/test/java/com/chesire/nekome/services/RefreshSeriesWorkerTests.kt
@@ -4,7 +4,7 @@ import android.content.Context
 import androidx.work.ListenableWorker
 import androidx.work.WorkerParameters
 import com.chesire.nekome.account.UserRepository
-import com.chesire.nekome.series.SeriesRepository
+import com.chesire.nekome.library.SeriesRepository
 import com.chesire.nekome.server.Resource
 import io.mockk.coEvery
 import io.mockk.coVerify
