+++ features/discover/build.gradle
@@ -31,7 +31,7 @@ dependencies {
     implementation fileTree(dir: "libs", include: ["*.jar"])
 
     implementation project(path: ":libraries:core")
-    implementation project(path: ":libraries:series")
+    implementation project(path: ':libraries:library')
     implementation project(path: ":libraries:server")
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
