+++ features/search/src/main/java/com/chesire/nekome/app/search/results/ResultsViewModel.kt
@@ -7,7 +7,7 @@ import com.chesire.nekome.core.AuthCaster
 import com.chesire.nekome.core.flags.SeriesType
 import com.chesire.nekome.core.models.SeriesModel
 import com.chesire.nekome.core.settings.ApplicationSettings
-import com.chesire.nekome.series.SeriesRepository
+import com.chesire.nekome.library.SeriesRepository
 import com.chesire.nekome.server.Resource
 import kotlinx.coroutines.launch
 import javax.inject.Inject
