+++ features/series/build.gradle
@@ -34,7 +34,7 @@ dependencies {
 
     implementation project(path: ":libraries:account")
     implementation project(path: ":libraries:core")
-    implementation project(path: ":libraries:series")
+    implementation project(path: ':libraries:library')
     implementation project(path: ":libraries:server")
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
