+++ features/series/src/main/java/com/chesire/nekome/app/series/list/SeriesListViewModel.kt
@@ -10,7 +10,7 @@ import com.chesire.nekome.core.extensions.postSuccess
 import com.chesire.nekome.core.flags.AsyncState
 import com.chesire.nekome.core.flags.UserSeriesStatus
 import com.chesire.nekome.core.models.SeriesModel
-import com.chesire.nekome.series.SeriesRepository
+import com.chesire.nekome.library.SeriesRepository
 import com.chesire.nekome.server.Resource
 import com.hadilq.liveevent.LiveEvent
 import kotlinx.coroutines.launch
