+++ features/series/src/test/java/com/chesire/nekome/app/series/detail/SeriesDetailViewModelTests.kt
@@ -4,7 +4,7 @@ import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.lifecycle.Observer
 import com.chesire.nekome.core.AuthCaster
 import com.chesire.nekome.core.flags.AsyncState
-import com.chesire.nekome.series.SeriesRepository
+import com.chesire.nekome.library.SeriesRepository
 import com.chesire.nekome.server.Resource
 import com.chesire.nekome.testing.CoroutinesMainDispatcherRule
 import com.chesire.nekome.testing.createSeriesModel
