+++ libraries/library/.gitignore
