+++ libraries/library/proguard-rules.pro
